requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
infobip-api-python-sdk==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==1.10.17
regex==2022.3.15
requests==2.31.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
python-dotenv==1.0.1
